+++ index.html
@@ -233,7 +233,6 @@
         <h2 class="data-error__title">Не удалось загрузить данные</h2>
       </section>
     </template>
-    <script src="js/function.js" defer></script>
-    <script src="js/main.js" defer></script>
+    <script src="js/main.js" type="module" defer></script>
 </body>
 </html>
